+++ lib/spec_forge/normalizer.rb
@@ -1,235 +1,19 @@
 # frozen_string_literal: true
 
+require_relative "normalizer/default"
+require_relative "normalizer/definition"
+require_relative "normalizer/normalization"
+require_relative "normalizer/validators"
+
 module SpecForge
-  #
-  # Base class for normalizing various data structures in SpecForge
-  #
-  # The Normalizer validates and standardizes user input from YAML files,
-  # ensuring it meets the expected structure and types before processing.
-  # It supports default values, type checking, aliases, and nested structures.
-  #
-  # @example Normalizing a hash
-  #   normalizer = Normalizer.new("spec", input_hash)
-  #   output, errors = normalizer.normalize
-  #
   class Normalizer
-    STRUCTURE = {
-      type: {
-        type: [String, Array],
-        default: nil,
-        validator: :present?
-      },
-      label: {
-        type: String,
-        default: nil
-      },
-      default: {
-        type: [String, NilClass, Numeric, Array, Hash, TrueClass, FalseClass],
-        default: nil
-      },
-      aliases: {
-        type: Array,
-        default: [],
-        structure: {type: String}
-      },
-      structure: {
-        type: Hash,
-        default: {}
-      },
-      validator: {
-        type: String,
-        default: nil
-      },
-      reference: {
-        type: String,
-        default: nil,
-        validator: :present?
-      }
-    }.freeze
-
     class << self
-      def default(structure_name: nil, structure: nil)
-        structure ||= @structures[structure_name]
-
-        if !structure.is_a?(Hash)
-          raise ArgumentError, "Invalid structure, provide either 'structure_name' or 'structure'"
-        end
-
-        default_type_value = lambda do |type_class|
-          case type_class
-          when Integer
-            0
-          when Proc
-            -> {}
-          else
-            type_class.new
-          end
-        end
-
-        structure.each_with_object({}) do |(key, value), hash|
-          type = value[:type]
-
-          hash[key] =
-            if value.key?(:default)
-              default = value[:default]
-              next if default.nil?
-
-              default.dup
-            elsif type.instance_of?(Array)
-              default_type_value.call(type.first)
-            else
-              default_type_value.call(type)
-            end
-        end
-      end
+      include Default
+      include Definition
+      include Normalization
 
       def normalize!(name, input, label: self.label)
         raise_errors! { normalize(name, input, label) }
-      end
-
-      #
-      # @api private
-      #
-      def normalize(input, structure_name:, label: self.label)
-        raise Error::InvalidTypeError.new(input, Hash, for: label) if !Type.hash?(input)
-
-        structure = @structures[structure_name]
-        if structure.nil?
-          structures = @structures.keys.to_or_sentence
-
-          raise ArgumentError,
-            "Invalid structure name. Expected #{structures}, got #{structure_name&.in_quotes}"
-        end
-
-        new(label, input, structure:).normalize
-      end
-
-      #
-      # Raises any errors collected by the block
-      #
-      # @yield Block that returns [output, errors]
-      # @yieldreturn [Array<Object, Set>] The result and any errors
-      #
-      # @return [Object] The normalized output if successful
-      #
-      # @raise [Error::InvalidStructureError] If any errors were encountered
-      #
-      # @api private
-      #
-      def raise_errors!(&block)
-        errors = Set.new
-
-        begin
-          output, new_errors = yield
-          errors.merge(new_errors) if new_errors.size > 0
-        rescue => e
-          errors << e
-        end
-
-        raise Error::InvalidStructureError.new(errors) if errors.size > 0
-
-        output
-      end
-
-      #
-      # @api private
-      #
-      def define
-        @normalizers = load_from_files
-        structures = @normalizers.delete(:_shared).stringify_keys
-
-        @normalizers.each do |normalizer_name, structure|
-          structure.transform_values!.with_key do |attribute, key|
-            case attribute
-            when String
-              hash = default(structure: STRUCTURE)
-              hash[:type] = resolve_type(attribute)
-              hash
-            when Hash
-              # A reference replaces the entire hash
-              if (name = attribute[:reference])
-                attribute = structures[name] || @normalizers[name]
-
-                if attribute.nil?
-                  structures_names = (structures.keys + @normalizers.keys)
-                    .map(&:in_quotes)
-                    .to_or_sentence
-
-                  raise Error, "Invalid reference name. Got #{name&.in_quotes}, expected one of #{structures_names}"
-                end
-              end
-
-              hash = raise_errors! do
-                new(
-                  "#{key.in_quotes} in normalizer/#{normalizer_name}.yml",
-                  attribute,
-                  structure: STRUCTURE
-                ).normalize
-              end
-
-              hash[:type] = resolve_type(hash[:type])
-              hash
-            else
-              raise ArgumentError,
-                "Invalid normalizer attribute. Expected String or Hash, got #{attribute.inspect}"
-            end
-          end
-        end
-      end
-
-      def resolve_type(type)
-        if type == "boolean"
-          [TrueClass, FalseClass]
-        elsif type.instance_of?(Array)
-          type.map { |t| resolve_type(t) }
-        else
-          type.classify.constantize
-        end
-      end
-
-      #
-      # @api private
-      #
-      def load_from_files
-        base_path = Pathname.new(File.expand_path("normalizers", __dir__))
-        paths = Dir[base_path.join("**/*.yml")].sort
-
-        paths.each_with_object({}) do |path, hash|
-          name = Pathname.new(path).relative_path_from(base_path).basename(".yml").to_s.to_sym
-          hash[name] = YAML.safe_load_file(path, symbolize_names: true)
-        end
-      end
-
-      #
-      # Defines the standard normalizer methods for a normalizer class
-      #
-      # This method creates three methods on the Normalizer class for a given normalizer:
-      # - default_#{key} - Returns default values for this normalizer
-      # - normalize_#{key}! - Normalize with error handling
-      # - normalize_#{key} - Normalize without error handling
-      #
-      # @param normalizer_class [Class] The normalizer class to define methods for
-      #
-      # @example Defining methods for a spec normalizer
-      #   define_normalizer_methods(SpecForge::Normalizer::Spec)
-      #   # Creates methods: default_spec, normalize_spec!, normalize_spec
-      #
-      # @private
-      #
-      def define_normalizer_methods(normalizer_class)
-        name = normalizer_class.id
-
-        Normalizer.define_singleton_method(:"default_#{name}") do
-          normalizer_class.default
-        end
-
-        Normalizer.define_singleton_method(:"normalize_#{name}!") do |input, **args|
-          normalizer_class.normalize!(input, **args)
-        end
-
-        Normalizer.define_singleton_method(:"normalize_#{name}") do |input, **args|
-          normalizer_class.normalize(input, **args)
-        end
       end
     end
 
@@ -467,5 +251,3 @@ module SpecForge
     define
   end
 end
-
-require_relative "normalizer/validators"
