+++ lib/spec_forge/normalizers/expectation.yml
@@ -1,0 +1,29 @@
+# Internal
+id:
+  reference: id
+line_number:
+  reference: line_number
+
+# User defined
+name:
+  reference: name
+base_url:
+  reference: base_url
+url:
+  reference: url
+http_verb:
+  reference: http_verb
+headers:
+  reference: headers
+query:
+  reference: query
+body:
+  reference: body
+variables:
+  reference: variables
+debug:
+  reference: debug
+store_as:
+  type: string
+  default: ""
+expect: hash
