+++ lib/spec_forge/normalizers/spec.yml
@@ -1,0 +1,28 @@
+# Internal
+id:
+  reference: id
+name:
+  reference: name
+file_name: string
+file_path: string
+line_number:
+  reference: line_number
+
+# User defined
+base_url:
+  reference: base_url
+url:
+  reference: url
+http_verb:
+  reference: http_verb
+headers:
+  reference: headers
+query:
+  reference: query
+body:
+  reference: body
+variables:
+  reference: variables
+debug:
+  reference: debug
+expectations: array
